Master with memory record
