+++ Radio-Compare/BconV2-Master/platformio.ini
@@ -12,5 +12,5 @@
 platform = atmelavr
 board = megaatmega2560
 framework = arduino
-upload_port = COM5
+upload_port = COM16
 monitor_speed = 115200
