+++ Radio-Compare/BconV2-Master/src/Settings.h
@@ -12,9 +12,10 @@
 #define Flashpin         21				// Flash memory SS pin
 uint16_t expectedDeviceID = 0xEF40;
 uint32_t Last_Address = 0;
+/// ///////
 String Msg2Write;
-uint8_t Stations = 5;
-
+uint8_t Stations = 6;
+uint32_t MyID = 0;
 //*******  Setup hardware pin definitions here ! ***************
 
 //These are the pin definitions for one of my own boards, the Easy Pro Mini,
