+++ Radio-Compare/BconV2-Slave/platformio.ini
@@ -12,5 +12,5 @@
 platform = atmelavr
 board = megaatmega2560
 framework = arduino
-upload_port =  COM6
+upload_port =  COM16
 monitor_speed = 115200
